requirements.txt
fastapi
pydantic
uvicorn[standard]
mysql-connector-python
requests
BeautifulSoup4